Move inspect elements to features/inspector
