+++ public/app/features/dashboard/components/Inspector/PanelInspector.tsx
@@ -2,13 +2,13 @@ import React, { useState } from 'react';
 import { connect, MapStateToProps } from 'react-redux';
 import { DashboardModel, PanelModel } from 'app/features/dashboard/state';
 import { PanelPlugin } from '@grafana/data';
-import { InspectTab } from '@grafana/ui';
 import { locationService } from '@grafana/runtime';
 
 import { StoreState } from 'app/types';
 import { GetDataOptions } from '../../../query/state/PanelQueryRunner';
 import { usePanelLatestData } from '../PanelEditor/usePanelLatestData';
 import { InspectContent } from './InspectContent';
+import { InspectTab } from 'app/features/inspector/types';
 import { useDatasourceMetadata, useInspectTabs } from './hooks';
 import { useLocation } from 'react-router-dom';
