+++ public/app/features/inspector/DetailText.tsx
