+++ public/app/features/inspector/InspectMetadataTab.tsx
