+++ public/app/features/inspector/types.ts
